rename TypesMigration -> Visibility::Migration
