+++ spec/graphql/query/result_spec.rb
@@ -25,7 +25,7 @@ describe GraphQL::Query::Result do
   it "exposes the context" do
     assert_instance_of GraphQL::Query::Context, result.context
     if GraphQL::Schema.use_schema_visibility?
-      assert_equal({a: :b, types_migration_running: true}, result.context.to_h)
+      assert_equal({a: :b, visibility_migration_running: true}, result.context.to_h)
     else
       assert_equal({a: :b}, result.context.to_h)
     end
